move referenceData out of lib
