+++ lib/index.js
@@ -2,7 +2,8 @@ import QUnitAdapter from "./adapters/QUnitAdapter.js";
 import JasmineAdapter from "./adapters/JasmineAdapter.js";
 import TapReporter from "./reporters/TapReporter.js";
 import ConsoleReporter from "./reporters/ConsoleReporter.js";
-import TestReporter, {testDataJasmine, testDataQUnit} from "./reporters/TestReporter.js";
+import TestReporter from "./reporters/TestReporter.js";
+import {Test, Suite} from "./Data.js";
 
 
 export default {
@@ -11,8 +12,6 @@ export default {
     TapReporter: TapReporter,
     ConsoleReporter: ConsoleReporter,
     TestReporter: TestReporter,
-    testData: {
-        QUnit: testDataQUnit,
-        Jasmine: testDataJasmine
-    }
+    Test: Test,
+    Suite: Suite
 };
