+++ test/test.js
@@ -1,5 +1,6 @@
 var Jasmine = require('jasmine');
 var JsReporters = require('../dist/js-reporters.js');
+var referenceData = require('./referenceData.js');
 
 var jasmine = new Jasmine();
 jasmine.loadConfig({
@@ -10,8 +11,7 @@ jasmine.loadConfig({
 });
 
 var runner =  new JsReporters.JasmineAdapter(jasmine);
-var referenceData = JsReporters.testData.Jasmine;
-var testReporter = new JsReporters.TestReporter(runner, referenceData );
+var testReporter = new JsReporters.TestReporter(runner, referenceData.Jasmine);
 
 jasmine.execute();
 if (!testReporter.ok){
